Add a new input device: Remote server
If a launchpad is not connected at startup, Lights is now launched
in server mode. In server mode, a ZeroMQ server listens on the local
port 7050. The server subscribes to the topic 'Colors' and accepts a
json encoded list of colors to be displayed.
The message formatting is the following:

Colors [{"blue": 0, "green": 0, "red": 255}, {"blue": 255, "green": 0, "red": 0}]

Given this message, the LED striped will alternate between red and blue.