Add a bi-directionnal ZeroMQ (Pub/Sub) Midi server
This change adds a Midi server to communicate with remote Midi devices.
It is based on ZeroMQ and a Publisher/Subscriber pattern.
Both the Server and the (soon to be) Client act simultaneously as a publisher
and a subscriber.
Communication from Client to Server happen under the topic 'MidiIn'.
Communication from Server to Client happen under the topic 'MidiOut'.